+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.8.0 - 2026-07-31]
+## [0.8.0 - 2026-08-06]
 ### ADDED
 - Add new ModifierRegistry to intercept recipes and modify them before running
 - Add `filelock` as dependency and include it's use in fetchez.core to prevent multiple processes from trying to fetch the same data.